Split apart hls files
